requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
autopep8==1.7.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
frozenlist==1.3.1
greenlet==1.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.5.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
orjson==3.8.0
priority==2.0.0
psycopg2-binary==2.9.3
pyaes==1.6.1
pycodestyle==2.9.1
pydantic==1.10.2
Pyrogram==2.0.57
PySocks==1.7.1
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quart==0.18.2
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tabulate==0.8.10
TgCrypto==1.2.3
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
